make the compiler more functional
Replace approach of having macros expand to macros and so on with a
simple function (translate) that translates from a JS syntax tree to a
CL form. Clean up scope handling.

This is not a finished change at this point, just a proof-of-concept!
Eval is broken, the arguments variable probably also in some
situations (!arguments is disabled, even though some stdlib stuff uses
it.)