Remove reader macro, add !include
Removes the #{javascript} reader macro, which was part of a vision of
mixing JS and CL that didn't work out. Add !include macro, which
allows a file of JavaScript to be translated and embedded into a CL
file, so that it will be part of the fasl file produced, and won't
have to be recompiled when the system is reloaded.