implement non-clos inline-cache system, clean up stdlib
this is not a finished commit -- some tests still fail, and there
are significant TODOs scattered all over the code

implements (in jsos.lisp) a system of implicit classes and inline
caching. also adds slot flags (ro, active, enumerable).

overhauls and cleans up the standard library implementation. this is
not quite finished yet (need to read the standard)